Datalake: Optimize writing for parquet
1. Optimize writing for parquet decimal.
2. Enable dictionary encoding for parquet.
3. Fixed the issue where writing decimals as fixed-length byte arrays resulted in excessively long outputs.

+++ contrib/datalake_fdw/src/datalake_def.h
@@ -283,7 +283,7 @@ typedef struct dataLakeOptions
 	/* parser filePath get prefix */
 	char		*prefix;
 	bool		readFdw;
-	size_t		fileSizeLimit;
+	int64_t		fileSizeLimit;
 	bool		vectorization;
 	/* iceberg & hudi options */
 	char		*table_identifier;
